feat: remove DluhyGraf cog and move functionality to a group within the Dluhy cog
